moved docs of the staging modules inside staging folder as well fixed imports names fixed proto files path
