+++ docs/developers/perform-transactions.md
@@ -6,12 +6,12 @@ As seeing inside the [FAQs](developer-faq.md#what-is-a-transaction), transaction
 Here is the list of currently available [messages](developer-faq.md#what-is-a-message) that you can use while creating transactions for the Desmos chain. 
 
 ### Posts
-* [`MsgCreatePost`](msgs/posts/create-post.md): allows you to create a new post or a comment for an existing post. 
-* [`MsgEditPost`](msgs/posts/edit-post.md): allows you to edit a previously created post message.
-* [`MsgAddPostReaction`](msgs/posts/add-post-reaction.md): allows you to add a reaction to an existing post. 
-* [`MsgRemovePostReaction`](msgs/posts/remove-post-reaction.md): allows you to remove a reaction from a post.
-* [`MsgAnswerPoll`](msgs/posts/answer-poll.md): allows you to answer a post's poll.
-* [`MsgRegisterReaction`](msgs/posts/register-reaction.md): allows you to register a reaction.
+* [`MsgCreatePost`](msgs/staging/posts/create-post.md): allows you to create a new post or a comment for an existing post. 
+* [`MsgEditPost`](msgs/staging/posts/edit-post.md): allows you to edit a previously created post message.
+* [`MsgAddPostReaction`](msgs/staging/posts/add-post-reaction.md): allows you to add a reaction to an existing post. 
+* [`MsgRemovePostReaction`](msgs/staging/posts/remove-post-reaction.md): allows you to remove a reaction from a post.
+* [`MsgAnswerPoll`](msgs/staging/posts/answer-poll.md): allows you to answer a post's poll.
+* [`MsgRegisterReaction`](msgs/staging/posts/register-reaction.md): allows you to register a reaction.
 
 ### Profiles
 * [`MsgSaveProfile`](msgs/profiles/save-profile.md): allows you to create or edit an existing profile.
@@ -22,13 +22,13 @@ Here is the list of currently available [messages](developer-faq.md#what-is-a-me
 * [`MsgCancelDTagTransferRequest`](msgs/profiles/cancel-dtag-transfer-request.md): allows the `dTag` request's sender to cancel his request.
 
 ### Relationships
-* [`MsgCreateRelationship`](msgs/relationships/create-relationship.md): allows you to create a relationship.
-* [`MsgDeleteRelationship`](msgs/relationships/delete-relationship.md): allows you to delete a relationship.
-* [`MsgBlockUser`](msgs/relationships/block-user.md): allows you to block a user.
-* [`MsgUnblockUser`](msgs/relationships/unblock-user.md): allows you to unblock a user.
+* [`MsgCreateRelationship`](msgs/staging/relationships/create-relationship.md): allows you to create a relationship.
+* [`MsgDeleteRelationship`](msgs/staging/relationships/delete-relationship.md): allows you to delete a relationship.
+* [`MsgBlockUser`](msgs/staging/relationships/block-user.md): allows you to block a user.
+* [`MsgUnblockUser`](msgs/staging/relationships/unblock-user.md): allows you to unblock a user.
 
 ### Reports
-* [`MsgReportPost`](msgs/reports/report-post.md): allows you to report an existing post.
+* [`MsgReportPost`](msgs/staging/reports/report-post.md): allows you to report an existing post.
 
 ### Params
-* [`EditParamsProposal`](msgs/edit_param_proposal.md): allows you to open a proposal to change profile's params.
+* [`EditParamsProposal`](msgs/staging/edit_param_proposal.md): allows you to open a proposal to change profile's params.
