+++ docs/developers/query-data.md
@@ -3,20 +3,20 @@
 Aside from [performing transactions](perform-transactions.md) you can also query for the current chain state in order to read the data stored inside it. Here you will find a list of all the query endpoints that are currently available.
 
 ## Posts
-- [Query the stored posts](queries/posts.md)
-- [Query the post's poll answers](queries/poll-answers.md)
-- [Query registered reactions](queries/reactions.md)
+- [Query the stored posts](queries/staging/posts.md)
+- [Query the post's poll answers](queries/staging/poll-answers.md)
+- [Query registered reactions](queries/staging/reactions.md)
 
 ## Profiles
 - [Query the stored profiles](queries/profile.md)
-- [Query all the DTag transfer requests](queries/user-dtag-requests.md)
+- [Query all the DTag transfer requests](queries/staging/user-dtag-requests.md)
 
 ## Relationships
-- [Query user's relationships](queries/user-relationships.md)
-- [Query user's blocked users](queries/user-blocks.md)
+- [Query user's relationships](queries/staging/user-relationships.md)
+- [Query user's blocked users](queries/staging/user-blocks.md)
 
 ## Reports
-- [Query the post's related reports](queries/reports.md)
+- [Query the post's related reports](queries/staging/reports.md)
 
 ## Modules Parameters
-- [Query parameters](queries/params.md)
+- [Query parameters](queries/staging/params.md)
