+++ proto/desmos/fees/v1beta1/genesis.proto
@@ -4,7 +4,7 @@ package desmos.fees.v1beta1;
 import "gogoproto/gogo.proto";
 import "desmos/fees/v1beta1/params.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/fees/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/fees/types";
 
 // GenesisState contains the data of the genesis state for the fees module
 message GenesisState {
