+++ proto/desmos/fees/v1beta1/params.proto
@@ -4,7 +4,7 @@ package desmos.fees.v1beta1;
 import "gogoproto/gogo.proto";
 import "desmos/fees/v1beta1/min_fee.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/fees/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/fees/types";
 
 // Params contains the parameters for the fees module
 message Params {
