+++ proto/desmos/fees/v1beta1/query.proto
@@ -5,7 +5,7 @@ import "gogoproto/gogo.proto";
 import "google/api/annotations.proto";
 import "desmos/fees/v1beta1/params.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/fees/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/fees/types";
 
 // Query defines the gRPC querier service.
 service Query {
