--- proto/desmos/magpie/v1beta1/genesis.proto
@@ -1,16 +1,0 @@
-syntax = "proto3";
-package desmos.magpie.v1beta1;
-
-import "gogoproto/gogo.proto";
-import "desmos/magpie/v1beta1/models.proto";
-
-option go_package = "github.com/desmos-labs/desmos/x/magpie/types";
-
-// GenesisState defines the magpie module's genesis state.
-message GenesisState {
-  uint64 default_session_length = 1
-      [ (gogoproto.moretags) = "yaml:\"default_session_length\"" ];
-
-  repeated desmos.magpie.v1beta1.Session sessions = 2
-      [ (gogoproto.nullable) = false ];
-}
