+++ proto/desmos/posts/v1beta1/genesis.proto
@@ -7,7 +7,7 @@ import "desmos/posts/v1beta1/polls.proto";
 import "desmos/posts/v1beta1/reactions.proto";
 import "desmos/posts/v1beta1/params.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/posts/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/posts/types";
 
 // GenesisState contains the data of the genesis state for the posts module
 message GenesisState {
