+++ proto/desmos/posts/v1beta1/params.proto
@@ -3,7 +3,7 @@ package desmos.posts.v1beta1;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/posts/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/posts/types";
 
 // Params contains the parameters for the posts module
 message Params {
