+++ proto/desmos/posts/v1beta1/polls.proto
@@ -4,7 +4,7 @@ package desmos.posts.v1beta1;
 import "gogoproto/gogo.proto";
 import "google/protobuf/timestamp.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/posts/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/posts/types";
 
 // PollAnswer contains the data of a single poll answer inserted by the creator
 message PollAnswer {
