+++ proto/desmos/profiles/v1beta1/genesis.proto
@@ -5,7 +5,7 @@ import "gogoproto/gogo.proto";
 import "desmos/profiles/v1beta1/params.proto";
 import "desmos/profiles/v1beta1/models.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/profiles/types";
 
 // GenesisState defines the profiles module's genesis state.
 message GenesisState {
