+++ proto/desmos/profiles/v1beta1/keeper/types.proto
@@ -4,7 +4,7 @@ package desmos.profiles.v1beta1.keeper;
 import "gogoproto/gogo.proto";
 import "desmos/profiles/v1beta1/models.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/profiles/keeper";
+option go_package = "github.com/desmos-labs/desmos/x/staging/profiles/keeper";
 
 // DTagOwner represents the owner of a specific DTag
 message WrappedDTagOwner {
