+++ proto/desmos/profiles/v1beta1/params.proto
@@ -3,7 +3,7 @@ package desmos.profiles.v1beta1;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/profiles/types";
 
 // Params contains the parameters for the profiles module
 message Params {
