+++ proto/desmos/profiles/v1beta1/query.proto
@@ -6,7 +6,7 @@ import "google/api/annotations.proto";
 import "desmos/profiles/v1beta1/params.proto";
 import "desmos/profiles/v1beta1/models.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/profiles/types";
 
 // Query defines the gRPC querier service.
 service Query {
