+++ proto/desmos/relationships/v1beta1/genesis.proto
@@ -4,7 +4,7 @@ package desmos.relationships.v1beta1;
 import "gogoproto/gogo.proto";
 import "desmos/relationships/v1beta1/models.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/relationships/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/relationships/types";
 
 // GenesisState defines the relationships module's genesis state.
 message GenesisState {
