+++ proto/desmos/relationships/v1beta1/models.proto
@@ -3,7 +3,7 @@ package desmos.relationships.v1beta1;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/relationships/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/relationships/types";
 
 // Relationship is the struct of a relationship.
 // It represent the concept of "follow" of traditional social networks.
