+++ proto/desmos/relationships/v1beta1/msgs.proto
@@ -3,7 +3,7 @@ package desmos.relationships.v1beta1;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/relationships/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/relationships/types";
 
 // Msg defines the relationships Msg service.
 service Msg {
