+++ proto/desmos/reports/v1beta1/genesis.proto
@@ -4,7 +4,7 @@ package desmos.reports.v1beta1;
 import "gogoproto/gogo.proto";
 import "desmos/reports/v1beta1/models.proto";
 
-option go_package = "github.com/desmos-labs/desmos/x/reports/types";
+option go_package = "github.com/desmos-labs/desmos/x/staging/reports/types";
 
 // GenesisState defines the reports module's genesis state.
 message GenesisState {
